Renamed dist folder to release
The dist folder is used by netbeans as the result of complete build, and be used as the distribution. This was in conflict with creating a new project with existing sources, so I renamed that folder to 'release', to mean that was what was released at the end of the project for evaluation.